requirements.txt
Flask-OAuthlib
Flask-PyMongo
Flask-Login
bcrypt
pyOpenSSL
redis
